requirements.txt
streamlit
pycryptodome
pillow
